Updated Mobile Pop-Up
