+++ src/components/Exams/Exams.js
@@ -1,7 +1,7 @@
 import {React, useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import { Mobile } from "../Mobile/Mobile";
-import PopUp from "../PopUp/PopUp";
+import PopUp from "../PopUp/DataPopUp";
 
 // API endpoint for fetching ALL exam data:
 // https://czi-covid-lypkrzry4q-uc.a.run.app/api/exams
